Extract the duration parsing into a util
